requirements.txt
# Core Data Processing
polars>=0.20.0
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0

# Database
duckdb>=0.10.0

# ETL & Orchestration
prefect>=2.14.0

# Data Quality & Profiling
great-expectations>=0.18.0
ydata-profiling>=4.6.0

# Visualization & Dashboards
streamlit>=1.30.0
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.2.0
shap>=0.44.0

# GenAI & RAG
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
google-generativeai>=0.3.0

# Optimization
scipy>=1.11.0
pulp>=2.7.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Streaming / Simulation
kafka-python>=2.0.2
watchdog>=3.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
